Resolve linker errors
